fix: split github-merge.lib.mjs to keep files under 1500 line limit
Move getLinkedPRsFromTimeline, syncReadyTags, and READY_LABEL to
github-merge-ready-sync.lib.mjs to reduce github-merge.lib.mjs from
1557 to 1325 lines (under the 1500 line CI limit).

Move Issue #1413 tests to test-timeline-linked-prs.mjs to reduce
test-merge-queue.mjs from 1547 to 1337 lines.

All exports are re-exported from github-merge.lib.mjs to preserve
backward compatibility for all importers.
